# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для записи к специалистам",
        "title": "LAPS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/schedules": {
            "get": {
                "description": "Возвращает recurring расписание специалиста (шаблон для всех недель)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Расписание"
                ],
                "summary": "Получить список расписаний",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID специалиста",
                        "name": "specialist_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "День недели (1=понедельник, 7=воскресенье)",
                        "name": "day_of_week",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Лимит (по умолчанию 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение (по умолчанию 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Расписание в формате недельного расписания",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет расписание специалиста на неделю",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Расписание"
                ],
                "summary": "Обновить расписание",
                "parameters": [
                    {
                        "description": "Данные для обновления расписания",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateScheduleDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успешном обновлении",
                        "schema": {
                            "$ref": "#/definitions/rest.messageResponseType"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает новое расписание для специалиста",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Расписание"
                ],
                "summary": "Создать расписание",
                "parameters": [
                    {
                        "description": "Данные для создания расписания",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateScheduleDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ID созданного расписания",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            }
        },
        "/schedules/free-slots": {
            "get": {
                "description": "Возвращает список временных слотов на выбранную дату с информацией о доступности",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Расписание"
                ],
                "summary": "Получить слоты специалиста",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID специалиста",
                        "name": "specialist_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Дата (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Вернуть детальную информацию (со всеми слотами и статусом)",
                        "name": "detailed",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список слотов",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            }
        },
        "/schedules/week": {
            "get": {
                "description": "Возвращает расписание специалиста на неделю в структурированном виде",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Расписание"
                ],
                "summary": "Получить недельное расписание специалиста",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID специалиста",
                        "name": "specialist_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Начало недели (YYYY-MM-DD), если не указано - текущая неделя",
                        "name": "week_start",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Недельное расписание",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            }
        },
        "/schedules/{id}": {
            "get": {
                "description": "Получает расписание по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Расписание"
                ],
                "summary": "Получить расписание по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID расписания",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Расписание",
                        "schema": {
                            "$ref": "#/definitions/domain.Schedule"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Расписание не найдено",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет существующее расписание",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Расписание"
                ],
                "summary": "Удалить расписание",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID расписания",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успешном удалении",
                        "schema": {
                            "$ref": "#/definitions/rest.messageResponseType"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Расписание не найдено",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.CreateScheduleDTO": {
            "type": "object",
            "required": [
                "slot_time",
                "week_schedule"
            ],
            "properties": {
                "slot_time": {
                    "type": "integer"
                },
                "week_schedule": {
                    "$ref": "#/definitions/domain.WeekSchedule"
                }
            }
        },
...
        "domain.DaySchedule": {
            "type": "object",
            "properties": {
                "work_time": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.WorkTimeSlot"
                    }
                }
            }
        },
...
        "domain.Schedule": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "day_of_week": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "exclude_times": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "slot_time": {
                    "type": "integer"
                },
                "specialist_id": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "domain.UpdateScheduleDTO": {
            "type": "object",
            "required": [
                "week_schedule"
            ],
            "properties": {
                "slot_time": {
                    "type": "integer"
                },
                "week_schedule": {
                    "$ref": "#/definitions/domain.WeekSchedule"
                }
            }
        },
...
        "domain.WeekSchedule": {
            "type": "object",
            "properties": {
                "friday": {
                    "$ref": "#/definitions/domain.DaySchedule"
                },
                "monday": {
                    "$ref": "#/definitions/domain.DaySchedule"
                },
                "saturday": {
                    "$ref": "#/definitions/domain.DaySchedule"
                },
                "sunday": {
                    "$ref": "#/definitions/domain.DaySchedule"
                },
                "thursday": {
                    "$ref": "#/definitions/domain.DaySchedule"
                },
                "tuesday": {
                    "$ref": "#/definitions/domain.DaySchedule"
                },
                "wednesday": {
                    "$ref": "#/definitions/domain.DaySchedule"
                }
            }
        },
...
        "domain.WorkTimeSlot": {
            "type": "object",
            "required": [
                "end_time",
                "start_time"
            ],
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "rest.errorResponseBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "rest.messageResponseType": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}